+++ desktop-build/service-manager/config.go
@@ -60,3 +60,4 @@ func ensureValidEnv() error {
 
 
 
+